Extract MethodThrowPointHelper to share throw point logic
Refactor duplicate throw point resolution from CastStringHandler,
MethodCallHandler, and StaticCallHandler into a shared helper class.
This also fixes StaticCallHandler missing the never-return-type
fallback to Throwable that MethodCallHandler already had.
