+++ src/Analyser/ExprHandler/CastStringHandler.php
@@ -4,23 +4,20 @@ namespace PHPStan\Analyser\ExprHandler;
 
 use PhpParser\Node\Expr;
 use PhpParser\Node\Expr\Cast;
+use PhpParser\Node\Identifier;
 use PhpParser\Node\Stmt;
 use PHPStan\Analyser\ExpressionContext;
 use PHPStan\Analyser\ExpressionResult;
 use PHPStan\Analyser\ExpressionResultStorage;
 use PHPStan\Analyser\ExprHandler;
+use PHPStan\Analyser\ExprHandler\Helper\MethodThrowPointHelper;
 use PHPStan\Analyser\ImpurePoint;
-use PHPStan\Analyser\InternalThrowPoint;
 use PHPStan\Analyser\MutatingScope;
 use PHPStan\Analyser\NodeScopeResolver;
-use PHPStan\DependencyInjection\AutowiredParameter;
 use PHPStan\DependencyInjection\AutowiredService;
 use PHPStan\Php\PhpVersion;
 use PHPStan\Reflection\InitializerExprTypeResolver;
-use PHPStan\Type\NeverType;
-use PHPStan\Type\ObjectType;
 use PHPStan\Type\Type;
-use Throwable;
 use function sprintf;
 
 /**
@@ -33,8 +30,7 @@ final class CastStringHandler implements ExprHandler
 	public function __construct(
 		private InitializerExprTypeResolver $initializerExprTypeResolver,
 		private PhpVersion $phpVersion,
-		#[AutowiredParameter(ref: '%exceptions.implicitThrows%')]
-		private bool $implicitThrows,
+		private MethodThrowPointHelper $methodThrowPointHelper,
 	)
 	{
 	}
@@ -64,20 +60,14 @@ final class CastStringHandler implements ExprHandler
 			}
 
 			if ($this->phpVersion->throwsOnStringCast()) {
-				$throwType = $toStringMethod->getThrowType();
-				if ($throwType === null) {
-					$returnType = $toStringMethod->getOnlyVariant()->getReturnType();
-					if ($returnType instanceof NeverType && $returnType->isExplicit()) {
-						$throwType = new ObjectType(Throwable::class);
-					}
-				}
-
-				if ($throwType !== null) {
-					if (!$throwType->isVoid()->yes()) {
-						$throwPoints[] = InternalThrowPoint::createExplicit($scope, $throwType, $expr, true);
-					}
-				} elseif ($this->implicitThrows) {
-					$throwPoints[] = InternalThrowPoint::createImplicit($scope, $expr);
+				$throwPoint = $this->methodThrowPointHelper->getThrowPoint(
+					$toStringMethod,
+					$toStringMethod->getOnlyVariant(),
+					new Expr\MethodCall($expr->expr, new Identifier('__toString')),
+					$scope
+				);
+				if ($throwPoint !== null) {
+					$throwPoints[] = $throwPoint;
 				}
 			}
 		}
